% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_count.R
\name{base_count}
\alias{base_count}
\title{Count bases in a DNA/RNA sequence}
\usage{
base_count(seq)
}
\arguments{
\item{seq}{character string with DNA or RNA sequence}
}
\value{
table with counts per base
}
\description{
Count bases in a DNA/RNA sequence
}
