% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_motif.R
\name{find_motif}
\alias{find_motif}
\title{Finds a pattern (motif) in a DNA/RNA sequence}
\usage{
find_motif(seq, motif)
}
\arguments{
\item{seq}{character string with DNA sequence}

\item{motif}{pattern to be found in the sequence}
}
\value{
list of positions where the pattern was found
}
\description{
Finds a pattern (motif) in a DNA/RNA sequence
}
