% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_orfs.R
\name{find_orfs}
\alias{find_orfs}
\title{Find ORFs in a sequence}
\usage{
find_orfs(seq)
}
\arguments{
\item{seq}{character string with RNA sequence}
}
\value{
table with frame and length of ORFs found
}
\description{
Find ORFs in a sequence
}
\details{
An ORF needs a start codon and an end codon.
This function considers 3 frames (0,1,2).
}
