% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gc_content.R
\name{gc_content}
\alias{gc_content}
\title{Calculate GC content of a DNA sequence}
\usage{
gc_content(seq)
}
\arguments{
\item{seq}{character string with DNA sequence}
}
\value{
numeric percentage of GC content
}
\description{
Calculate GC content of a DNA sequence
}
