% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hamming_distance.R
\name{hamming_distance}
\alias{hamming_distance}
\title{Finds differences in two sequences with same length}
\usage{
hamming_distance(seq1, seq2)
}
\arguments{
\item{seq1}{character string with DNA/RNA sequence}

\item{seq2}{character string with DNA/RNA sequence}
}
\value{
number of differences
}
\description{
Finds differences in two sequences with same length
}
\details{
Both sequences must have equal length for this function.
}
