% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/melting_temp.R
\name{melting_temp}
\alias{melting_temp}
\title{Calculate melting temperature}
\usage{
melting_temp(seq)
}
\arguments{
\item{seq}{character string with DNA sequence}
}
\value{
melting temperature in °C
}
\description{
Calculate melting temperature
}
\details{
This function uses Wallace rule to calculate melting temperature
Calculation: 4(g+c)+2(a+t)
}
