% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_complement.R
\name{reverse_complement}
\alias{reverse_complement}
\title{Gets reverse complement of a DNA sequence}
\usage{
reverse_complement(seq)
}
\arguments{
\item{seq}{character string with DNA sequence}
}
\value{
character string with reverse complement
}
\description{
Gets reverse complement of a DNA sequence
}
