% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reverse_transcribe.R
\name{reverse_transcribe}
\alias{reverse_transcribe}
\title{Converts RNA to DNA by replacing U with T}
\usage{
reverse_transcribe(seq)
}
\arguments{
\item{seq}{character string with RNA sequence}
}
\value{
character string with reverse-transcribed sequence
}
\description{
Converts RNA to DNA by replacing U with T
}
