% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transcribe.R
\name{transcribe}
\alias{transcribe}
\title{Converts DNA to RNA by replacing T with U}
\usage{
transcribe(seq)
}
\arguments{
\item{seq}{character string with DNA sequence}
}
\value{
character string with transcribed sequence
}
\description{
Converts DNA to RNA by replacing T with U
}
