% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_orfs.R
\name{translate_orfs}
\alias{translate_orfs}
\title{Translate ORFs in a sequence}
\usage{
translate_orfs(seq)
}
\arguments{
\item{seq}{character string with RNA sequence}
}
\value{
aminoacids sequence from ORFs in RNA input
}
\description{
Translate ORFs in a sequence
}
\details{
The sequence needs a start codon and an end codon to translate.
This function considers 3 frames (0,1,2).
}
